Inflate as helper in noir wasm and fix browser test
